% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_lightgbm}
\alias{install_lightgbm}
\title{Installs LightGBM}
\usage{
install_lightgbm(...)
}
\arguments{
\item{...}{additional parameters passed to \code{\link[utils:install.packages]{utils::install.packages()}}}
}
\description{
Installs LightGBM
}
